requirements.txt
asgiref==3.5.0
Django==3.2
django-debug-toolbar==3.2.4
django-extensions==3.1.5
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
split-settings==1.0.0
sqlparse==0.4.2
Werkzeug==2.0.3
